newRepOpenAI/requirements.txt
-e ./texasholdem    
openai>=1.0.0
python-dotenv>=1.0.0 
maturin>=0.16.0
transformers>=4.45.2
torch>=2.5.0
torchvision>=0.18.0
torchaudio>=2.5.0
accelerate
pandas>=1.5.0